% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{change_labeles_to_factors}
\alias{change_labeles_to_factors}
\title{Change labelled data to factors}
\usage{
change_labeles_to_factors(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with labelled columns converted to factors
}
\description{
Converts all \link{haven}-labelled columns to factors.
}
\seealso{
Other data-prep: 
\code{\link{build_imp_column_names}()},
\code{\link{build_sat_column_names}()},
\code{\link{convert_labels_to_row_names}()},
\code{\link{prep_data}()},
\code{\link{remove_data_prefix}()},
\code{\link{remove_data_suffix}()},
\code{\link{replace_spaces_with_underscores}()}
}
\concept{data-prep}
