% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{convert_labels_to_row_names}
\alias{convert_labels_to_row_names}
\title{Convert labels to column names}
\usage{
convert_labels_to_row_names(df)
}
\arguments{
\item{df}{A data frame with SPSS labels}
}
\value{
A data frame with labels as column names
}
\description{
Swaps SPSS variable labels into column names using \code{\link[sjlabelled:label_to_colnames]{sjlabelled::label_to_colnames()}}.
}
\seealso{
Other data-prep: 
\code{\link{build_imp_column_names}()},
\code{\link{build_sat_column_names}()},
\code{\link{change_labeles_to_factors}()},
\code{\link{prep_data}()},
\code{\link{remove_data_prefix}()},
\code{\link{remove_data_suffix}()},
\code{\link{replace_spaces_with_underscores}()}
}
\concept{data-prep}
