% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.normalized_scores}
\alias{get.normalized_scores}
\title{Get normalized scores}
\usage{
get.normalized_scores(your_df)
}
\arguments{
\item{your_df}{A data frame containing JTBD scores}
}
\value{
A data frame with normalized scores (0-1 range)
}
\description{
Min-max normalizes scores within each segment and measure.
}
