% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_count}
\alias{get_count}
\title{Get count of factor levels}
\usage{
get_count(data_frame, col_name)
}
\arguments{
\item{data_frame}{A data frame containing the column to be counted}

\item{col_name}{The name of the column to count}
}
\value{
A data frame with counts for each factor level
}
\description{
Get count of factor levels
}
