% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_segment.comp}
\alias{get_jtbd_segment.comp}
\title{Get JTBD segment comparison}
\usage{
get_jtbd_segment.comp(your_df)
}
\arguments{
\item{your_df}{A data frame containing JTBD scores}
}
\value{
A data frame with segment comparison results
}
\description{
Get JTBD segment comparison
}
