% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_var_values.list.300}
\alias{get_jtbd_var_values.list.300}
\title{Get JTBD variable values list (300+ occurrences)}
\usage{
get_jtbd_var_values.list.300(data_frame, segmentation_column)
}
\arguments{
\item{data_frame}{A data frame containing the segmentation column}

\item{segmentation_column}{The name of the segmentation column}
}
\value{
A list of unique values from the segmentation column with more than 300 occurrences
}
\description{
Get JTBD variable values list (300+ occurrences)
}
