% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_min_max}
\alias{get_min_max}
\title{Get min and max values for JTBD scores}
\usage{
get_min_max(your_long_data_frame, job_steps = NULL)
}
\arguments{
\item{your_long_data_frame}{A long format data frame}

\item{job_steps}{Character vector of job steps to filter (default: all)}
}
\value{
A data frame with min, max, and difference calculations
}
\description{
Get min and max values for JTBD scores
}
