% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{jtbd_colors}
\alias{jtbd_colors}
\title{JTBD color palette}
\usage{
jtbd_colors()
}
\value{
A named character vector of colors
}
\description{
Returns a named color palette for JTBD opportunity classifications.
}
\examples{
jtbd_colors()
}
