% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtbdtools-package.R
\docType{package}
\name{jtbdtools-package}
\alias{jtbdtools}
\alias{jtbdtools-package}
\title{jtbdtools: Quantitative Jobs-to-Be-Done Analysis Using Outcome-Driven Innovation}
\description{
Implements Ulwick's Outcome-Driven Innovation opportunity scoring formula. Calculates importance, satisfaction, and opportunity scores from Likert-scale survey data. Provides ggplot2 visualizations and gt tables for segment comparison and opportunity identification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlesrogers/jtbdtools}
  \item \url{https://charlesrogers.github.io/jtbdtools/}
  \item Report bugs at \url{https://github.com/charlesrogers/jtbdtools/issues}
}

}
\author{
\strong{Maintainer}: Charles Rogers \email{charlesrogers@gmail.com} (\href{https://orcid.org/0009-0004-3162-0322}{ORCID})

}
\keyword{internal}
