% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.job_step}
\alias{plot.job_step}
\title{Plot job step scatter}
\usage{
\method{plot}{job_step}(x, step_title = "", ...)
}
\arguments{
\item{x}{A data frame containing job step data}

\item{step_title}{The title for the plot}

\item{...}{Additional arguments (unused)}
}
\value{
A ggplot object
}
\description{
Creates an importance vs satisfaction scatter plot for a job step.
}
\seealso{
Other visualization: 
\code{\link{plot_cleveland}()},
\code{\link{plot_opportunity_matrix}()},
\code{\link{plot_this.graph.abs_score}()},
\code{\link{plot_this.graph.rel.abs_score}()},
\code{\link{plot_this.graph.rel_score}()},
\code{\link{theme_jtbd}()}
}
\concept{visualization}
