% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cleveland}
\alias{plot_cleveland}
\title{Cleveland (lollipop) comparison plot}
\usage{
plot_cleveland(
  data,
  objective,
  group_1,
  group_2,
  title_string,
  subtitle_string
)
}
\arguments{
\item{data}{A data frame with objective and segment score columns}

\item{objective}{Unquoted column name for objectives}

\item{group_1}{Unquoted column name for first segment scores}

\item{group_2}{Unquoted column name for second segment scores}

\item{title_string}{Plot title}

\item{subtitle_string}{Plot subtitle}
}
\value{
A ggplot object
}
\description{
Compares two segments side-by-side using a Cleveland dot plot.
Ported from the quantjtbd package.
}
\seealso{
Other visualization: 
\code{\link{plot.job_step}()},
\code{\link{plot_opportunity_matrix}()},
\code{\link{plot_this.graph.abs_score}()},
\code{\link{plot_this.graph.rel.abs_score}()},
\code{\link{plot_this.graph.rel_score}()},
\code{\link{theme_jtbd}()}
}
\concept{visualization}
