% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{plot_segment_index}
\alias{plot_segment_index}
\title{Plot segment index heatmap}
\usage{
plot_segment_index(
  profile_result,
  max_vars = 6,
  title = "Segment Index: Over & Under-Represented Traits",
  n = NULL,
  study = NULL
)
}
\arguments{
\item{profile_result}{Result from \code{\link[=jtbd_profile_segments]{jtbd_profile_segments()}}}

\item{max_vars}{Maximum number of variables to show (default: 6)}

\item{title}{Plot title}

\item{n, study}{Sample size and study label for the plot footer (see \code{\link[=jtbd_footer]{jtbd_footer()}}).}
}
\value{
A ggplot object
}
\description{
Shows how each segment over- or under-indexes on profiling attributes
relative to the overall population. Index of 100 = same as population.
}
\examples{
data(jtbd_sample)
cl <- jtbd_cluster(jtbd_sample, n_clusters = 3)
prof <- jtbd_profile_segments(cl$data)
plot_segment_index(prof)
}
\seealso{
Other clustering: 
\code{\link{create_persona_table}()},
\code{\link{jtbd_cluster}()},
\code{\link{jtbd_cluster_profile}()},
\code{\link{jtbd_feature_matrix}()},
\code{\link{jtbd_find_k}()},
\code{\link{jtbd_pca}()},
\code{\link{jtbd_profile_segments}()},
\code{\link{jtbd_segment}()},
\code{\link{plot_cluster_heatmap}()},
\code{\link{plot_elbow}()},
\code{\link{plot_pca_biplot}()},
\code{\link{plot_pca_loadings}()},
\code{\link{plot_pca_scree}()},
\code{\link{plot_segment_dna}()},
\code{\link{plot_segment_fingerprint}()},
\code{\link{plot_segment_profiles}()},
\code{\link{plot_segment_radar}()},
\code{\link{plot_segment_radar_facet}()}
}
\concept{clustering}
