% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_this.graph.rel.abs_score}
\alias{plot_this.graph.rel.abs_score}
\title{Plot relative and absolute score graphs}
\usage{
plot_this.graph.rel.abs_score(
  your_data_frame,
  seg_value_title,
  last_value,
  save_path = tempdir()
)
}
\arguments{
\item{your_data_frame}{A data frame from \code{\link[=get_jtbd_segment.comp.ordinal]{get_jtbd_segment.comp.ordinal()}}}

\item{seg_value_title}{Title label for the segment variable}

\item{last_value}{The segment value to label on the right side of the chart}

\item{save_path}{Directory to save plots (default: \code{\link[=tempdir]{tempdir()}})}
}
\value{
NULL (called for side effects: generates and saves plots)
}
\description{
Generates both relative (percent-of-segment-max) and absolute score bump charts
for segment comparison. Convenience wrapper around \code{\link[=plot_this.graph.rel_score]{plot_this.graph.rel_score()}}
and \code{\link[=plot_this.graph.abs_score]{plot_this.graph.abs_score()}}.
}
\seealso{
Other visualization: 
\code{\link{plot.job_step}()},
\code{\link{plot_cleveland}()},
\code{\link{plot_opportunity_matrix}()},
\code{\link{plot_this.graph.abs_score}()},
\code{\link{plot_this.graph.rel_score}()},
\code{\link{theme_jtbd}()}
}
\concept{visualization}
