% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{remove_data_suffix}
\alias{remove_data_suffix}
\title{Remove data suffix}
\usage{
remove_data_suffix(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with suffixes removed from column names
}
\description{
Removes everything after and including \code{-} from column names.
}
\seealso{
Other data-prep: 
\code{\link{build_imp_column_names}()},
\code{\link{build_sat_column_names}()},
\code{\link{change_labeles_to_factors}()},
\code{\link{convert_labels_to_row_names}()},
\code{\link{prep_data}()},
\code{\link{remove_data_prefix}()},
\code{\link{replace_spaces_with_underscores}()}
}
\concept{data-prep}
