% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{replace_spaces_with_underscores}
\alias{replace_spaces_with_underscores}
\title{Replace spaces with underscores in column names}
\usage{
replace_spaces_with_underscores(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with cleaned column names
}
\description{
Uses \code{\link[janitor:clean_names]{janitor::clean_names()}} to standardize column names.
}
\seealso{
Other data-prep: 
\code{\link{build_imp_column_names}()},
\code{\link{build_sat_column_names}()},
\code{\link{change_labeles_to_factors}()},
\code{\link{convert_labels_to_row_names}()},
\code{\link{prep_data}()},
\code{\link{remove_data_prefix}()},
\code{\link{remove_data_suffix}()}
}
\concept{data-prep}
