% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{build_imp_column_names}
\alias{build_imp_column_names}
\title{Build importance column names}
\usage{
build_imp_column_names(df, job_section)
}
\arguments{
\item{df}{A data frame to process}

\item{job_section}{The job section to add to the column names}
}
\value{
A data frame with updated column names
}
\description{
Build importance column names
}
