% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{calculate_opportunity_score}
\alias{calculate_opportunity_score}
\title{Calculate opportunity score}
\usage{
calculate_opportunity_score(data_frame_split)
}
\arguments{
\item{data_frame_split}{A data frame with split importance and satisfaction columns}
}
\value{
A data frame with calculated opportunity scores
}
\description{
Calculate opportunity score
}
