% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{calculate_pop_pct_score}
\alias{calculate_pop_pct_score}
\title{Calculate population percentage score}
\usage{
calculate_pop_pct_score(objectives)
}
\arguments{
\item{objectives}{A data frame containing objectives to be scored}
}
\value{
A data frame with calculated scores for each objective
}
\description{
Calculate population percentage score
}
