% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{change_labeles_to_factors}
\alias{change_labeles_to_factors}
\title{Change labelled data to factors}
\usage{
change_labeles_to_factors(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with labelled columns converted to factors
}
\description{
Change labelled data to factors
}
