% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{convert_labels_to_row_names}
\alias{convert_labels_to_row_names}
\title{Convert labels to row names}
\usage{
convert_labels_to_row_names(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with labels converted to column names
}
\description{
Convert labels to row names
}
