% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{create.job_step.table}
\alias{create.job_step.table}
\title{Create job step table}
\usage{
create.job_step.table(df, job.string, job_step.string, path.table)
}
\arguments{
\item{df}{A data frame containing job step data}

\item{job.string}{The job string}

\item{job_step.string}{The job step string}

\item{path.table}{The path to save the table}
}
\value{
Saves a png file of the job step table
}
\description{
Create job step table
}
