% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{create.pct.table}
\alias{create.pct.table}
\title{Create percentage table}
\usage{
create.pct.table(df, job.string, var_name.string, job_step.string, path.table)
}
\arguments{
\item{df}{Data frame containing the data}

\item{job.string}{Job string}

\item{var_name.string}{Variable name string}

\item{job_step.string}{Job step string}

\item{path.table}{Path to save the table}
}
\value{
Saves a PNG file of the percentage table
}
\description{
Create percentage table
}
