% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{find_imp_sat_columns}
\alias{find_imp_sat_columns}
\title{Find importance and satisfaction columns}
\usage{
find_imp_sat_columns(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing importance and satisfaction columns}
}
\value{
A data frame with only importance and satisfaction columns
}
\description{
Extracts all columns starting with \code{imp_} or \code{sat_} from a data frame.
}
\seealso{
Other scoring: 
\code{\link{calculate_opportunity_score}()},
\code{\link{calculate_pop_pct_score}()},
\code{\link{get_jtbd_scores}()},
\code{\link{get_jtbd_scores.individual}()},
\code{\link{get_sample_size}()},
\code{\link{split_imp_sat_columns}()}
}
\concept{scoring}
