% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{find_imp_sat_columns}
\alias{find_imp_sat_columns}
\title{Find importance and satisfaction columns}
\usage{
find_imp_sat_columns(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing importance and satisfaction columns}
}
\value{
A data frame with only importance and satisfaction columns
}
\description{
Find importance and satisfaction columns
}
