% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get.normalized_scores}
\alias{get.normalized_scores}
\title{Get normalized scores}
\usage{
get.normalized_scores(your_df)
}
\arguments{
\item{your_df}{A data frame containing JTBD scores}
}
\value{
A data frame with normalized scores
}
\description{
Get normalized scores
}
