% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get.percent_of_max}
\alias{get.percent_of_max}
\title{Get percent of max for each function}
\usage{
get.percent_of_max(your_df)
}
\arguments{
\item{your_df}{A data frame containing JTBD scores}
}
\value{
A data frame with percent of max calculations
}
\description{
Get percent of max for each function
}
