% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_scores.batch}
\alias{get_jtbd_scores.batch}
\title{Calculate JTBD scores for multiple segments}
\usage{
get_jtbd_scores.batch(
  master_table,
  data_frame,
  merged_df_and_segmentation_column,
  static_segment,
  list_of_unique_segments
)
}
\arguments{
\item{master_table}{The main data frame to update}

\item{data_frame}{The data frame containing JTBD data}

\item{merged_df_and_segmentation_column}{The column used for segmentation}

\item{static_segment}{The static segment to compare against}

\item{list_of_unique_segments}{A list of unique segments to analyze}
}
\value{
An updated master table with JTBD scores for each segment
}
\description{
Calculate JTBD scores for multiple segments
}
