% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_scores.individual}
\alias{get_jtbd_scores.individual}
\title{Calculate individual JTBD scores}
\usage{
get_jtbd_scores.individual(df.uid_imp_sat)
}
\arguments{
\item{df.uid_imp_sat}{A data frame containing importance and satisfaction scores with user IDs}
}
\value{
A data frame with individual JTBD scores
}
\description{
Calculate individual JTBD scores
}
