% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_scores.pairwise}
\alias{get_jtbd_scores.pairwise}
\title{Calculate JTBD scores for a pair of segments}
\usage{
get_jtbd_scores.pairwise(
  your_data_frame,
  column_to_split_on,
  factor_a,
  factor_b
)
}
\arguments{
\item{your_data_frame}{The data frame containing JTBD data}

\item{column_to_split_on}{The column used for segmentation}

\item{factor_a}{The first segment to compare}

\item{factor_b}{The second segment to compare}
}
\value{
A data frame with JTBD scores for the two segments
}
\description{
Calculate JTBD scores for a pair of segments
}
