% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_scores}
\alias{get_jtbd_scores}
\title{Calculate JTBD scores}
\usage{
get_jtbd_scores(your_data_frame, segment_var)
}
\arguments{
\item{your_data_frame}{A data frame containing importance and satisfaction data}

\item{segment_var}{The segmentation variable}
}
\value{
A data frame with calculated JTBD scores
}
\description{
Calculate JTBD scores
}
