% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_segment.comp.ordinal}
\alias{get_jtbd_segment.comp.ordinal}
\title{Get JTBD segment comparison: ordinal}
\usage{
get_jtbd_segment.comp.ordinal(your_df)
}
\arguments{
\item{your_df}{A data frame containing JTBD scores}
}
\value{
A data frame with ordinal comparison results
}
\description{
Get JTBD segment comparison: ordinal
}
