% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_segment.comp_and_plot}
\alias{get_jtbd_segment.comp_and_plot}
\title{Compare and plot JTBD segment scores}
\usage{
get_jtbd_segment.comp_and_plot(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing JTBD scores}
}
\value{
Plots and saves JTBD score comparisons
}
\description{
Compare and plot JTBD segment scores
}
