% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{get_jtbd_var_values.list}
\alias{get_jtbd_var_values.list}
\title{Get unique segment values}
\usage{
get_jtbd_var_values.list(data_frame, segmentation_column, min_n = 30)
}
\arguments{
\item{data_frame}{A data frame containing the segmentation column}

\item{segmentation_column}{The name of the segmentation column (as string)}

\item{min_n}{Minimum number of respondents to include a segment (default: 30)}
}
\value{
A character vector of segment values
}
\description{
Extracts unique values from a segmentation column, filtering to segments
with at least \code{min_n} respondents.
}
\examples{
data(jtbd_sample)
get_jtbd_var_values.list(jtbd_sample, "segment", min_n = 30)
}
\seealso{
Other segmentation: 
\code{\link{get_jtbd_scores.batch}()},
\code{\link{get_jtbd_scores.comparison}()},
\code{\link{get_jtbd_scores.pairwise}()},
\code{\link{get_jtbd_segment.comp.ordinal}()},
\code{\link{get_seg_comp.build.linear}()},
\code{\link{test_segment_significance}()}
}
\concept{segmentation}
