% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_jtbd_var_values.list}
\alias{get_jtbd_var_values.list}
\title{Get JTBD variable values list}
\usage{
get_jtbd_var_values.list(data_frame, segmentation_column)
}
\arguments{
\item{data_frame}{A data frame containing the segmentation column}

\item{segmentation_column}{The name of the segmentation column}
}
\value{
A list of unique values from the segmentation column with more than 90 occurrences
}
\description{
Get JTBD variable values list
}
