% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_min_max}
\alias{get_min_max}
\title{Get min and max values for JTBD scores}
\usage{
get_min_max(your_long_data_frame)
}
\arguments{
\item{your_long_data_frame}{A long format data frame containing JTBD scores}
}
\value{
A data frame with min, max, and difference calculations
}
\description{
Get min and max values for JTBD scores
}
