% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Get sample size}
\usage{
get_sample_size(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing columns starting with "imp__"}
}
\value{
Integer: number of non-NA rows
}
\description{
Returns the number of non-NA respondents based on the last importance column.
}
\examples{
data(jtbd_sample)
get_sample_size(jtbd_sample)
}
\seealso{
Other scoring: 
\code{\link{add_score_cis}()},
\code{\link{calculate_opportunity_score}()},
\code{\link{calculate_pop_pct_score}()},
\code{\link{find_imp_sat_columns}()},
\code{\link{get_jtbd_scores}()},
\code{\link{get_jtbd_scores.individual}()},
\code{\link{split_imp_sat_columns}()}
}
\concept{scoring}
