% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Get sample size}
\usage{
get_sample_size(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing columns starting with "imp__"}
}
\value{
The number of non-NA rows in the last column starting with "imp__"
}
\description{
Get sample size
}
