% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{get_seg_comp.build.linear}
\alias{get_seg_comp.build.linear}
\title{Build group score segmentation}
\usage{
get_seg_comp.build.linear(total_pop, original_df, new_table)
}
\arguments{
\item{total_pop}{Total population data frame}

\item{original_df}{Original data frame}

\item{new_table}{New table to be added}
}
\value{
An updated data frame with new segmentation scores
}
\description{
Build group score segmentation
}
