% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{make_data_long.pairwise}
\alias{make_data_long.pairwise}
\title{Convert JTBD scores to long format for pairwise comparison}
\usage{
make_data_long.pairwise(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing JTBD scores}
}
\value{
A long format data frame
}
\description{
Convert JTBD scores to long format for pairwise comparison
}
