% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{plot.job_step}
\alias{plot.job_step}
\title{Plot job step}
\usage{
\method{plot}{job_step}(your.df, step_title)
}
\arguments{
\item{your.df}{A data frame containing job step data}

\item{step_title}{The title for the job step}
}
\value{
A ggplot object with job step visualization
}
\description{
Plot job step
}
