% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Elbow plot for cluster evaluation}
\usage{
plot_elbow(k_results, n = NULL, study = NULL)
}
\arguments{
\item{k_results}{Result from \code{\link[=jtbd_find_k]{jtbd_find_k()}}}

\item{n, study}{Sample size and study label for the plot footer (see \code{\link[=jtbd_footer]{jtbd_footer()}}).}
}
\value{
A ggplot object
}
\description{
Shows within-cluster sum of squares (WCSS) and average silhouette width
for different numbers of clusters.
}
\examples{
data(jtbd_sample)
k_eval <- jtbd_find_k(jtbd_sample, max_k = 5)
plot_elbow(k_eval)
}
\seealso{
Other clustering: 
\code{\link{create_persona_table}()},
\code{\link{jtbd_cluster}()},
\code{\link{jtbd_cluster_profile}()},
\code{\link{jtbd_feature_matrix}()},
\code{\link{jtbd_find_k}()},
\code{\link{jtbd_pca}()},
\code{\link{jtbd_profile_segments}()},
\code{\link{jtbd_segment}()},
\code{\link{plot_cluster_heatmap}()},
\code{\link{plot_pca_biplot}()},
\code{\link{plot_pca_loadings}()},
\code{\link{plot_pca_scree}()},
\code{\link{plot_segment_dna}()},
\code{\link{plot_segment_fingerprint}()},
\code{\link{plot_segment_index}()},
\code{\link{plot_segment_profiles}()},
\code{\link{plot_segment_radar}()},
\code{\link{plot_segment_radar_facet}()}
}
\concept{clustering}
