% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca_biplot}
\alias{plot_pca_biplot}
\title{PCA biplot}
\usage{
plot_pca_biplot(pca_result, cluster_labels = NULL)
}
\arguments{
\item{pca_result}{Result from \code{\link[=jtbd_pca]{jtbd_pca()}}}

\item{cluster_labels}{Optional factor/integer vector of cluster assignments}
}
\value{
A ggplot object
}
\description{
2D scatter of respondents on PC1 vs PC2, optionally colored by cluster.
}
\examples{
data(jtbd_sample)
cl <- jtbd_cluster(jtbd_sample, n_clusters = 3)
plot_pca_biplot(cl$pca, cl$cluster)
}
\seealso{
Other clustering: 
\code{\link{create_persona_table}()},
\code{\link{jtbd_cluster}()},
\code{\link{jtbd_cluster_profile}()},
\code{\link{jtbd_feature_matrix}()},
\code{\link{jtbd_find_k}()},
\code{\link{jtbd_pca}()},
\code{\link{jtbd_profile_segments}()},
\code{\link{jtbd_segment}()},
\code{\link{plot_cluster_heatmap}()},
\code{\link{plot_elbow}()},
\code{\link{plot_pca_loadings}()},
\code{\link{plot_pca_scree}()},
\code{\link{plot_segment_dna}()},
\code{\link{plot_segment_fingerprint}()},
\code{\link{plot_segment_index}()},
\code{\link{plot_segment_profiles}()},
\code{\link{plot_segment_radar}()},
\code{\link{plot_segment_radar_facet}()}
}
\concept{clustering}
