% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca_loadings}
\alias{plot_pca_loadings}
\title{PCA loadings bar chart}
\usage{
plot_pca_loadings(pca_result, component = 1)
}
\arguments{
\item{pca_result}{Result from \code{\link[=jtbd_pca]{jtbd_pca()}}}

\item{component}{Which component to plot (default: 1)}
}
\value{
A ggplot object
}
\description{
Shows how each objective loads on a given principal component.
High-loading objectives define the "theme" of that component.
}
\seealso{
Other clustering: 
\code{\link{jtbd_cluster}()},
\code{\link{jtbd_cluster_profile}()},
\code{\link{jtbd_feature_matrix}()},
\code{\link{jtbd_find_k}()},
\code{\link{jtbd_pca}()},
\code{\link{jtbd_segment}()},
\code{\link{plot_cluster_heatmap}()},
\code{\link{plot_elbow}()},
\code{\link{plot_pca_biplot}()},
\code{\link{plot_pca_scree}()}
}
\concept{clustering}
