% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca_scree}
\alias{plot_pca_scree}
\title{PCA scree plot}
\usage{
plot_pca_scree(pca_result)
}
\arguments{
\item{pca_result}{Result from \code{\link[=jtbd_pca]{jtbd_pca()}}}
}
\value{
A ggplot object
}
\description{
Shows eigenvalues per component with a Kaiser line at 1.0 to help
determine how many components to retain.
}
\examples{
data(jtbd_sample)
pca <- jtbd_pca(jtbd_sample)
plot_pca_scree(pca)
}
\seealso{
Other clustering: 
\code{\link{create_persona_table}()},
\code{\link{jtbd_cluster}()},
\code{\link{jtbd_cluster_profile}()},
\code{\link{jtbd_feature_matrix}()},
\code{\link{jtbd_find_k}()},
\code{\link{jtbd_pca}()},
\code{\link{jtbd_profile_segments}()},
\code{\link{jtbd_segment}()},
\code{\link{plot_cluster_heatmap}()},
\code{\link{plot_elbow}()},
\code{\link{plot_pca_biplot}()},
\code{\link{plot_pca_loadings}()},
\code{\link{plot_segment_dna}()},
\code{\link{plot_segment_fingerprint}()},
\code{\link{plot_segment_index}()},
\code{\link{plot_segment_profiles}()},
\code{\link{plot_segment_radar}()},
\code{\link{plot_segment_radar_facet}()}
}
\concept{clustering}
