% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{plot_this.graph.rel_score}
\alias{plot_this.graph.rel_score}
\title{Plot relative score graph}
\usage{
plot_this.graph.rel_score(your_data_frame, seg_value_title, last_value)
}
\arguments{
\item{your_data_frame}{A data frame containing JTBD scores}

\item{seg_value_title}{The title for the segment value}

\item{last_value}{The last value to be labeled}
}
\value{
A ggplot object
}
\description{
Plot relative score graph
}
