% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{plot_this.group.plotable}
\alias{plot_this.group.plotable}
\title{Plot JTBD scores for group comparison}
\usage{
plot_this.group.plotable(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame containing JTBD scores}
}
\value{
A long format data frame suitable for plotting
}
\description{
Plot JTBD scores for group comparison
}
