% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for analysis}
\usage{
prep_data(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A processed data frame
}
\description{
Prepare data for analysis
}
