% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{remove_data_prefix}
\alias{remove_data_prefix}
\title{Remove data prefix}
\usage{
remove_data_prefix(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with prefixes removed from column names
}
\description{
Remove data prefix
}
