% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{remove_data_suffix}
\alias{remove_data_suffix}
\title{Remove data suffix}
\usage{
remove_data_suffix(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with suffixes removed from column names
}
\description{
Remove data suffix
}
