% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{remove_weird_text_formatting.jtbd}
\alias{remove_weird_text_formatting.jtbd}
\title{Remove weird text formatting for JTBD}
\usage{
remove_weird_text_formatting.jtbd(your_data_frame)
}
\arguments{
\item{your_data_frame}{A data frame to process}
}
\value{
A data frame with cleaned text formatting
}
\description{
Remove weird text formatting for JTBD
}
