% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{replace_spaces_with_underscores}
\alias{replace_spaces_with_underscores}
\title{Replace spaces with underscores in column names}
\usage{
replace_spaces_with_underscores(df)
}
\arguments{
\item{df}{A data frame to process}
}
\value{
A data frame with spaces replaced by underscores in column names
}
\description{
Replace spaces with underscores in column names
}
