% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{split_imp_sat_columns}
\alias{split_imp_sat_columns}
\title{Split importance and satisfaction columns}
\usage{
split_imp_sat_columns(data_frame_imp_sat)
}
\arguments{
\item{data_frame_imp_sat}{A data frame with an \code{objective_name} column}
}
\value{
A data frame with \code{imp_sat} and \code{objective} columns
}
\description{
Separates the \code{objective_name} column into \code{imp_sat} (imp/sat indicator) and
\code{objective} (the actual objective name) by splitting on \verb{__}.
}
\seealso{
Other scoring: 
\code{\link{add_score_cis}()},
\code{\link{calculate_opportunity_score}()},
\code{\link{calculate_pop_pct_score}()},
\code{\link{find_imp_sat_columns}()},
\code{\link{get_jtbd_scores}()},
\code{\link{get_jtbd_scores.individual}()},
\code{\link{get_sample_size}()}
}
\concept{scoring}
