% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{split_imp_sat_columns}
\alias{split_imp_sat_columns}
\title{Split importance and satisfaction columns}
\usage{
split_imp_sat_columns(data_frame_imp_sat)
}
\arguments{
\item{data_frame_imp_sat}{A data frame with importance and satisfaction columns}
}
\value{
A data frame with split importance and satisfaction columns
}
\description{
Split importance and satisfaction columns
}
