% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantjtbd-package.R
\name{theme.job_step}
\alias{theme.job_step}
\title{Theme for job step visualization}
\usage{
theme.job_step(your.df, step_title)
}
\arguments{
\item{your.df}{A data frame containing job step data}

\item{step_title}{The title for the job step}
}
\value{
A gt table with formatted job step data
}
\description{
Theme for job step visualization
}
